# pubky-homeserver/openapi.yml
openapi: 3.1.0
info:
  title: Pubky Homeserver API
  description: |
    The Pubky Homeserver exposes three servers:
    - **Client Server** — Primary API for user authentication, file storage, and event streaming.
    - **Admin Server** — Password-protected administration endpoints.
    - **Metrics Server** — Prometheus metrics (`GET /metrics`).

    This specification covers the Client Server and Admin Server.

    ## Tenant Identification

    Per-tenant (user) routes resolve the target user from these sources (highest priority first):
    1. `pubky-host` header (explicit override; takes precedence over `Host`)
    2. `Host` header (e.g. from TLS SNI on the Pubky socket)
    3. `?pubky-host=` query parameter (fallback when no key found in headers)

    The value must be a z-base-32 encoded Ed25519 public key.

    ## Authentication

    Two authentication methods exist:
    - **Grant-based (current):** `Authorization: Bearer <token>` obtained via `POST /auth/grant/session`.
    - **Cookie-based (deprecated):** Session cookie set by `POST /signup` or `POST /session`.

    Grant middleware runs first; if it produces a session, cookie middleware is skipped.

    ## Capabilities

    Write operations are gated by capability-based access control. Capabilities are
    path-scoped (e.g. `read_write:/pub/app/`) or `root` (full access to `/pub/`).

    ## Write Path Restriction

    All PUT/DELETE operations require paths under `/pub/`. Attempts to write elsewhere
...
  version: 0.9.0
...
servers:
...
paths:
...
  /generate_signup_token:
    get:
      tags: [Admin]
      summary: Generate signup token (default limits)
      operationId: generateSignupToken
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      responses:
        "200":
          description: Generated token
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Missing or invalid admin password

    post:
      tags: [Admin]
      summary: Generate signup token (custom limits)
      operationId: generateSignupTokenWithLimits
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserQuota"
            examples:
              customLimits:
                value:
                  storage_quota_mb: 1024
                  rate_read: "200mb/m"
              unlimited:
                value:
                  storage_quota_mb: "unlimited"
                  rate_read: "unlimited"
                  rate_write: "unlimited"
      responses:
        "200":
          description: Generated token
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Missing or invalid admin password
        "422":
          description: Invalid quota format
  /signup_tokens:
    get:
      tags: [Admin]
      summary: List signup tokens
      description: |
        Paginated list of signup tokens with usage information.
        Supports filtering by state (`all`, `used`, `unused`).
      operationId: listSignupTokens
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of tokens to return per page.
          schema:
            type: integer
            minimum: 1
            maximum: 65535
        - name: cursor
          in: query
          description: Pagination cursor (token string from `next_cursor` of a previous response).
          schema:
            type: string
        - name: state
          in: query
          description: Filter by token state.
          schema:
            type: string
            enum: [all, used, unused]
            default: all
      responses:
        "200":
          description: Paginated list of signup tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignupTokensResponse"
        "400":
          description: Invalid query parameters (bad cursor format, unknown state, limit=0)
        "401":
          description: Missing or invalid admin password
  /webdav/{entry_path}:
    delete:
      tags: [Admin]
      summary: Delete entry via admin
      operationId: adminDeleteEntry
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      parameters:
        - name: entry_path
          in: path
          required: true
          description: "Full path: `{pubkey}/path` (e.g. `o1gg96ewuo.../pub/file.txt`)"
          schema:
            type: string
      responses:
        "204":
          description: Entry deleted
        "400":
          description: Invalid pubkey format
        "401":
          description: Missing or invalid admin password
        "404":
          description: Entry not found
...
  /users/{pubkey}/disable:
    post:
      tags: [Admin]
      summary: Disable user account
      operationId: disableUser
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      parameters:
        - $ref: "#/components/parameters/AdminPubkey"
      responses:
        "200":
          description: User disabled
          content:
            text/plain:
              schema:
                type: string
                example: Ok
        "400":
          description: Invalid pubkey format
        "401":
          description: Missing or invalid admin password
        "404":
          description: User not found
  /users/{pubkey}/enable:
    post:
      tags: [Admin]
      summary: Enable user account
      operationId: enableUser
      servers:
        - url: http://localhost:6288
      security:
        - adminPassword: []
      parameters:
        - $ref: "#/components/parameters/AdminPubkey"
      responses:
        "200":
          description: User enabled
          content:
            text/plain:
              schema:
                type: string
                example: Ok
        "400":
          description: Invalid pubkey format
        "401":
          description: Missing or invalid admin password
        "404":
          description: User not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Grant-based bearer token from `POST /auth/grant/session`.

    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: |
        Deprecated. In practice, the cookie name is the user's z-base-32
        public key and the value is the session secret. The `name` field here
        is a placeholder; the actual name varies per user.

    adminPassword:
      type: apiKey
      in: header
      name: X-Admin-Password
      description: Admin server password.

    adminBasicAuth:
      type: http
      scheme: basic
      description: Basic auth for WebDAV endpoints. Username `admin`, password is the admin password.
  parameters:
...
    AdminPubkey:
      name: pubkey
      in: path
      required: true
      description: User's z-base-32 encoded public key.
      schema:
        type: string
  schemas:
...
    SignupTokensResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SignupTokenItem"
        next_cursor:
          type: ["string", "null"]
          description: Cursor for the next page, or `null` if this is the last page.
    SignupTokenItem:
      type: object
      required: [token, created_at, used_at, used_by]
      properties:
        token:
          type: string
          description: The signup token string.
        created_at:
          type: string
          format: date-time
          description: When the token was created.
        used_at:
          type: ["string", "null"]
          format: date-time
          description: When the token was used, or `null` if unused.
        used_by:
          type: ["string", "null"]
          description: z-base-32 public key of the user who used the token, or `null` if unused.

    # ── Admin ───────────────────────────────────────────────────────────
...
    UserQuota:
      type: object
      description: |
        Quota configuration. All fields are optional.
        - Absent or `null`: use system default
        - `"unlimited"`: no limit
        - Value: explicit limit
      properties:
        storage_quota_mb:
          description: Storage quota in megabytes, `"unlimited"`, or `null` for default.
          oneOf:
            - type: integer
            - type: string
              enum: [unlimited]
        rate_read:
          description: "Read bandwidth limit (e.g. `\"200mb/m\"`, `\"unlimited\"`, or `null`)."
          type: string
        rate_write:
          description: "Write bandwidth limit (e.g. `\"100mb/m\"`, `\"unlimited\"`, or `null`)."
          type: string
        rate_read_burst:
          description: |
            Burst override for read rate limit, in the rate's natural unit
            (MB for "…mb/s", KB for "…kb/s"). `null` or absent = burst equals rate (default).
          type: integer
          format: int32
          minimum: 1
        rate_write_burst:
          description: |
            Burst override for write rate limit, in the rate's natural unit
            (MB for "…mb/s", KB for "…kb/s"). `null` or absent = burst equals rate (default).
          type: integer
          format: int32
          minimum: 1
        allowed_write_paths:
          description: |
            Restrict which paths a user can write to.
            - `null` or absent: unrestricted (all paths allowed) — default
            - `[]`: read-only (no writes allowed)
            - `["/pub/tokens/", "/pub/paykit/"]`: only these prefix paths
          type: array
          items:
            type: string
...